refactor: Separate the package named strategy
