+++ src/main/java/ladder/domain/LineFactory.java
@@ -1,5 +1,7 @@
 package ladder.domain;
 
+import ladder.strategy.LineStrategy;
+
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
