+++ src/test/java/ladder/strategy/RandomLineStrategyTest.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.strategy;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
